### 2023-10-22 #5970 @editorialbot
[PRE REVIEW]: Machine Learning Validation via Rational Dataset Sampling with astartes
**Submitting author:** @JacksonBurns (<a href="http://orcid.org/0000-0002-0657-9426">Jackson Burns</a>)
**Repository:** https://github.com/JacksonBurns/astartes
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.1.3.post1
