### 2023-10-22 #5970 @arfon
Noting that this submission was reviewed by pyOpenSci here: https://github.com/pyOpenSci/software-submission/issues/120
