### 2023-10-22 #5970 @arfon
@lwasser – for rOpenSci submissions we've started trying to list the reviewers as the reviewers here too. Does that sound good to you @BerylKanali @du-phan (this will associate your GitHub handles with the JOSS paper).
