### 2023-10-23 #5970 @BerylKanali
> @lwasser – for rOpenSci submissions we've started trying to list the reviewers as the reviewers here too. We could do the same for pyOpenSci too. Does that sound good to you too @BerylKanali @du-phan (this will associate your GitHub handles with the JOSS paper).

@arfon Sounds good to me.
