### 2023-10-23 #5970 @lwasser
@arfon i second @BerylKanali reply - yes please do list our reviewers here. many thanks for recognizing them in the JOSS paper! Let's model your partnership with rOpenSci and make this a part of our pyOpenSci-JOSS partnership. If you need anything from me please say the word!
Many thanks.
