### 2023-10-28 #5996 @arfon
@JacksonBurns – I'll need to do a quick check of your paper but in the interim, could you please take a look at the DOI suggestions from @editorialbot and see if any of them are matches? If they are, please add them to your BibTeX file.
