### 2023-10-31 #5996 @JacksonBurns
> @JacksonBurns – I'll need to do a quick check of your paper but in the interim, could you please take a look at the DOI suggestions from @editorialbot and see if any of them are matches? If they are, please add them to your BibTeX file.

@arfon done! I have confirmed that the paper still compiles, as well.

Quite a handy bot! :robot:
