### 2023-11-05 #5996 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
